RecyclerView #4

+++ app/src/main/java/com/example/timerapp/database/PresetTimer.kt
@@ -2,17 +2,16 @@ package com.example.timerapp.database
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
-import androidx.room.PrimaryKey
 
 @Entity(primaryKeys = ["name", "presetName", "timerOrder"])
-data class PresetTimer (
+data class PresetTimer(
     val name: String,
     val presetName: String,
     val timerOrder: Int,
     @ColumnInfo
-    val presetTime: Int = 0,
+    val presetTime: Long = 0L,
     @ColumnInfo
-    val notificationTime: Int = 0,
+    val notificationTime: Long = 0L,
     @ColumnInfo
     val isSelected: Boolean = false
 )
